+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Log a count of the files to be spellchecked.
+
 ## [2.0.0] - 2017-12-28
 
 ### Added